+++ CHANGELOG.md
@@ -5,3 +5,4 @@ v0.1.1 - XXXX-XX-XX
 Added
 - New dispatch option `atomic_priority_set` to set priority and dispatch event in one atomic operation.
 - New dispatch option for `members` called `external` to dispatch event to nodes except the local node.
+- New dispatch option for `members` to dispatch event to list of selected nodes.